abstract out expr representation
